+++ src/api/objects/toot.ts
@@ -38,7 +38,7 @@ import {
     TootLike,
     TootScore,
     TrendingLink,
-    TrendingTag,
+    TagWithUsageCounts,
     WeightName
 } from "../../types";
 
@@ -82,7 +82,7 @@ export interface SerializableToot extends mastodon.v1.Status {
     sources?: string[];                // Source of the toot (e.g. trending tag toots, home timeline, etc.)
     trendingLinks?: TrendingLink[];    // Links that are trending in this toot
     trendingRank?: number;             // Most trending on a server gets a 10, next is a 9, etc.
-    trendingTags?: TrendingTag[];      // Tags that are trending in this toot
+    trendingTags?: TagWithUsageCounts[];      // Tags that are trending in this toot
     audioAttachments?: mastodon.v1.MediaAttachment[];
     imageAttachments?: mastodon.v1.MediaAttachment[];
     videoAttachments?: mastodon.v1.MediaAttachment[];
@@ -143,14 +143,14 @@ export default class Toot implements TootObj {
     // extensions to mastodon.v1.Status. Most of these are set in completeProperties()
     completedAt?: string;
     followedTags?: mastodon.v1.Tag[];            // Array of tags that the user follows that exist in this toot
-    participatedTags?: TrendingTag[];            // Array of tags that the user has participated in that exist in this toot
+    participatedTags?: TagWithUsageCounts[];            // Array of tags that the user has participated in that exist in this toot
     @Type(() => Account) reblogsBy!: Account[];  // The accounts that retooted this toot
     @Type(() => Toot) resolvedToot?: Toot;       // This Toot with URLs resolved to homeserver versions
     scoreInfo?: TootScore;                       // Scoring info for weighting/sorting this toot
     sources?: string[];                          // Source of the toot (e.g. trending tag toots, home timeline, etc.)
     trendingRank?: number;                       // Most trending on a server gets a 10, next is a 9, etc.
     trendingLinks?: TrendingLink[];              // Links that are trending in this toot
-    trendingTags?: TrendingTag[];                // Tags that are trending that appear in this toot
+    trendingTags?: TagWithUsageCounts[];                // Tags that are trending that appear in this toot
     audioAttachments!: mastodon.v1.MediaAttachment[];
     imageAttachments!: mastodon.v1.MediaAttachment[];
     videoAttachments!: mastodon.v1.MediaAttachment[];
@@ -502,7 +502,7 @@ export default class Toot implements TootObj {
     private completeProperties(
         userData: UserData,
         trendingLinks: TrendingLink[],
-        trendingTags: TrendingTag[],
+        trendingTags: TagWithUsageCounts[],
         isDeepInspect?: boolean
     ): void {
         if (!this.shouldComplete()) return;